requirement.txt
django==5.0.0
djangorestframework
pyyaml
requests
django-cors-headers
